table2screen fixed AND DFT improvements AND terminal control

+++ src/WaveOsc.hpp
@@ -1,7 +1,7 @@
 // WaveOsc.h
 #ifndef WAVEOSC_H
 #define WAVEOSC_H
-double getWavetableValue(double* wave_table);
+float getWavetableValue(double* wave_table);
 void setfreq(double f);
 #endif
